requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
tiktoken==0.7.0
